rename functions and delete user util file
